feat: 컴포넌트 분리 및 navbar 렌더링
